requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
psycopg2-binary
pillow~=10.2.0
pytils
python-dotenv~=1.0.1
pip~=23.2.1
pytz~=2024.1
requests~=2.31.0
setuptools~=65.5.0
djangorestframework~=3.14.0
django-filter
djangorestframework-simplejwt
coverage
drf-yasg-1.21.7
django-cors-headers-4.3.1
